% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetBinaryDosageInformation}
\alias{GetBinaryDosageInformation}
\title{Function to produce summary of a binary dosage file used by GxEScan}
\usage{
GetBinaryDosageInformation(binaryDosageFilename, nSub, nSNPs)
}
\arguments{
\item{nSub}{Number of subjects in the genetic data file. This is needed for formats 1, 2 and 3.
This value is returned for format 4 in the results}

\item{nSNPs}{Number of SNPs in the genetic data file. This is needed for formats 1, 2 and 3.
This value is returned for format 4 in the results}

\item{binaryDosageFilesname}{Name of file with genetic data, normally ends with .bdosage}
}
\value{
List with data needed by GxEScan
}
\description{
Function to produce summary of a binary dosage file used by GxEScan
}
