% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticData.R
\name{GetGeneticInfo}
\alias{GetGeneticInfo}
\title{Function read information about genetic data needed for GxEScan}
\usage{
GetGeneticInfo(geneticFile, familyFile, mapFile)
}
\arguments{
\item{geneticFile}{Name of file with genetic data - currently only binary dosage
files are supported.}

\item{familyFile}{Name of file with family data corresponding to the genetic data
file. This must be in plink format.}

\item{mapFile}{Name of the map file associated with the genetic data file. This
must be in extended plink map file format.}
}
\value{
List with information about genetic data need by the GxEScan routine
}
\description{
Function read information about genetic data needed for GxEScan
This information is then passed to GxEScan
}
