% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute2.R
\name{GetI2Info}
\alias{GetI2Info}
\title{Function to get information about a file in an Impute 2 format}
\usage{
GetI2Info(i2file, header = TRUE, snpCol = c("SNP", "CHR", "BP", "A1", "A2"),
  subID, startCol, format = 2L, usesFID = TRUE, sep = "\\t")
}
\arguments{
\item{i2file}{Impute 2 file name. Must be provided.}

\item{header}{Does the Impute 2 file have a header? Default value is TRUE.}

\item{snpCol}{Must be provided if there is no header.
If this case this must be a integer vector of the column numbers for SNP name,
chromosome, location in BP, reference allele, and alternate allele respectively.
All must be provided a value of 0 indicates the value isn't in the file. SNP name
column is required.
If provided and there is a header, the vector may integer vector as described
above or it may be a character array with the column names of the appropriate
values. Unused columns can be indicated the an empty string or NA. If snpCol
is not provided and there is a header, snpCol takes on the default values of
'SNP', 'CHR', 'BP', 'A1' and 'A2'.}

\item{subID}{A data frame with two columns. Each column must be a character. The first column
is the family ID and the second column is the subject ID. If family IDs are not
used the family ID column must be filled with empty strings and not NAs.}

\item{format}{Number of genetic probabilities in file, must be 1, 2 or 3
1 indicates that only the dosage is in the file Default value is 2.}

\item{usesFID}{Does the impute file contain both the family and subject ID? Not used if
there is not header. Default value is TRUE.}

\item{sep}{Separator used in file. Currently only space and tabs are allowed.
Default value is tab.}

\item{startcol}{Column number where dosages or genetic probabilities begin. Must be provided.}
}
\value{
List with information needed by GxEScan routine to read SNP data from file
}
\description{
Function to get information about a file in an Impute 2 like format.
An Impute 2 like format has a group of columns with SNP information
followed by a group of dosage or genetic probabilities for the subjects.
There must be a column with the SNP name that uniquely identifies the SNP.
Optional column values can include chromosome, location in base pairs,
reference and alternate alleles. The file may contain a header. If there
is a header, the columns with SNP data can be found using the column names
or by specifying the column numbers. If there is a header, the subject and
family IDs are read from the header. If there is not header, the subject
and family IDs must be specified.
}
