% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxeScanR.R
\name{gwis}
\alias{gwis}
\title{gwis}
\usage{
gwis(data, bdinfo, snps, outfile, skipfile, minmaf, blksize, binary)
}
\arguments{
\item{data}{Data frame containing the subject ID, phenotype
and covariates}

\item{bdinfo}{Information about the binary dosage file returned
from the BinaryDosage::getbdinfo routine}

\item{outfile}{The file name for the results output.}

\item{minmaf}{Minimum minor allele frequency of SNPs to include
in analysis. SNPS that have less than 20 minor alleles observed
will be excluded from the analysis regardless of the value of
minmaf. A value of 0 indicates to use all the SNPs that have 20
minor alleles observed. Default value is 0.}

\item{blksize}{Size of blocks of SNPs to read in at one time.
Larger blocks can improve overall speed but require larger
amounts of computer memory. A value of 0 indicates to use the
recommended block size. Default value is 0.}

\item{binary}{Logical value indicating if the phenotype
is a binary value. Default value is false.}
}
\value{
0
}
\description{
Run a gwas using genetic data from a binary dosage file
}
\examples{
data <- data.frame()
bdinfofile <- system.file("extdata", "bdinfo_set1a_1_1.rds", package = "bdgwas")
bdinfo <- readRDS(bdinfofile)
outfile <- "test.out"
gwas(data = data, bdinfo = bdinfo, outfile = outfile)
}
