% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEOnly.R
\name{GxEOnly}
\alias{GxEOnly}
\title{Function to fit the D|G,E and D|G,E,GxE models and return the results
for the G, GxE, and the two degree of freedom test for both}
\usage{
GxEOnly(subjectData, geneticData, outputFile, minMAF1 = 0.05,
  minMAF2 = 0.05, snps)
}
\arguments{
\item{subjectData}{Data frame with subject data
First column must be a character string
If second column is a character string, the first column is the
is the family ID and the second column is the subject ID otherwise
the first column is the subject
The column following the subject ID is the subject's phenotype
this column must be a numeric value and not an integer
The remian columns are the covariates values. These must be numeric
values. The covariate in the last column is the covariate interacting
with the gene.}

\item{geneticData}{List with information on reading genetic data
This is returned from GetGeneticInfo}

\item{outputFile}{Name of file to write the results to}

\item{minMAF1}{Minimum minor allele frequency in complete data set. Has to be a value between 0.0001 and 0.25.
Default value is 0.05.}

\item{minMAF2}{Minimum minor allele frequency in subjects with complete data. Has to be a value between 0.0001 and 0.25.
Default value is 0.05.}

\item{snps}{Subset of SNPs to use. This can be a character vector of SNP names or an
integer vector of locations in the snps data frame in the genetic data list}
}
\value{
0 - success
1 - failure
}
\description{
Function to fit the D|G,E and D|G,E,GxE models and return the results
for the G, GxE, and the two degree of freedom test for both
}
