% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEScan.R
\name{GxEScan}
\alias{GxEScan}
\title{Function to test format of subject data table passed to GxEScan}
\usage{
GxEScan(subjectData, geneticData, outputFile, skippedFilename = "",
  minMaf = 0.05, geCutoff = 0.001, snps)
}
\arguments{
\item{subjectData}{Data frame with subject data
First column must be a character string
If second column is a character string, the first column is the
is the family ID and the second column is the subject ID otherwise
the first column is the subject
The column following the subject ID is the subject's phenotype
this column must be a numeric value and not an integer
The remian columns are the covariates values. These must be numeric
values. The covariate in the last column is the covariate interacting
with the gene.}

\item{geneticData}{List with information on reading genetic data
This is returned from GetGeneticInfo}

\item{outputFile}{Name of file to write the results to}

\item{skippedFilename}{Name of file to write info about SNPs that were skipped. If this is blank
no file is written. If this is the same as outputFilename, the skipped SNPs
are written to the output file along with NA for all tests.}

\item{minMaf}{Minimum minor allele frequency. Has to be a value between 0.0001 and 0.25.
Default value is 0.05.}

\item{geCutoff}{p-value cut off for fitting polytomous logistic regression models given the results
from the the logistic model assuming Hardy-Weinberg equilibrium. Must be a
value fromm 0 to 1. Default value is 0.001.}

\item{snps}{Subset of SNPs to use. This can be a character vector of SNP names or an
integer vector of locations in the snps data frame in the genetic data list}
}
\value{
0 - success
1 - failure
}
\description{
Function to test format of subject data table passed to GxEScan
}
