% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GxEScanC}
\alias{GxEScanC}
\title{Function to fit models scanning over genotypes}
\usage{
GxEScanC(subjectData, geneticInfo)
}
\arguments{
\item{subjectData}{List returned from SubsetSubjects}

\item{geneticInfo}{List returned from one of the functions to get the required information
about the source of genetic data.}
}
\value{
0 success
1 failure
}
\description{
Function to fit selected models over genotypes. Results from
these models are then processed by GxETest to display the results
for these one step tests along with two step tests derived from
the one step tests.
}
