% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GxEScanCSubset}
\alias{GxEScanCSubset}
\title{Function to fit models scanning over a subset of genotypes}
\usage{
GxEScanCSubset(subjectData, geneticInfo, outputFilename, skippedFilename,
  minMaf, geCutoff, snpIndices)
}
\arguments{
\item{subjectData}{List returned from SubsetSubjects}

\item{geneticInfo}{List returned from one of the functions to get the required information
about the source of genetic data.}

\item{outputFilename}{Name of output file}

\item{skippedFilename}{Name of file to write info about SNPs that were skipped. If this is blank
no file is written. If this is the same as outputFilename, the skipped SNPs
are written to the output file along with NA for all tests.}

\item{minMaf}{Minimum minor allele frequency. Must be between 0.0001 and 0.25}

\item{geCutoff}{p-value cut off value for logistic G|E test to fit polytomous models.}

\item{snpIndices}{Indices of SNP locations in geneticInfo to be used in analysis}
}
\value{
0 success
1 failure
}
\description{
Function to fit selected models over a subset of genotypes. Results from
these models are then processed by GxETest to display the results
for these one step tests along with two step tests derived from
the one step tests.
}
