% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GxEScanFreqC2}
\alias{GxEScanFreqC2}
\title{Function to calculate allele frequencies}
\usage{
GxEScanFreqC2(subjectData, geneticInfo, outputFilename)
}
\arguments{
\item{subjectData}{List returned from SubsetSubjects}

\item{geneticInfo}{List returned from one of the functions to get the required information
about the source of genetic data.}

\item{outputFilename}{Name of output file}
}
\value{
0 success
1 failure
}
\description{
Function to calculate allele frequencies
}
