\name{GxEScanR-package}
\alias{GxEScanR-package}
\alias{GxEScanR}
\docType{package}
\title{
\packageTitle{GxEScanR}
}
\description{
\packageDescription{GxEScanR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GxEScanR}
\packageIndices{GxEScanR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GxEScanR}

Maintainer: \packageMaintainer{GxEScanR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
