% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GxETest}
\alias{GxETest}
\title{Function to display the results from the scans performed by GxEScan}
\usage{
GxETest()
}
\value{
0 success
1 failure
}
\description{
Function to display the results from the scans performed by GxEScan.
The results include Manhattan and qq-plot along with tables of the
top hits for each test. The results are saved to an Excel spreadsheet.
}
