% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imp2toBD.R
\name{Imp2toBD}
\alias{Imp2toBD}
\title{Function to create a binary dosage file from an Impute 2 style format.}
\usage{
Imp2toBD(i2info, filename, format = 1, subformat = 1)
}
\arguments{
\item{i2info}{Information about Impute 2 file. List returned from GetI2Info}

\item{filename}{Base name of binary dosage file to be written.
The binary dosage file will have the extension .bdose.
If format written is less 3 a family and map file will be created with
the family file having the extension .fam and the map file having the
extenstion .map}

\item{format}{Format of the binary dosage file to generate.}

\item{subformat}{Subformat of the binary dosage file to generate.}
}
\value{
0 - success, 1 - failure
}
\description{
Function to create a binary dosage file from an Impute 2 style format.
}
