% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Imp2toBDC}
\alias{Imp2toBDC}
\title{Function to convert and Impute 2 style file to a binary dosage file}
\usage{
Imp2toBDC(imp2Info, filename, format, subformat)
}
\arguments{
\item{imp2Info}{List containing information of Impute 2 file, returned from GetI2Info}

\item{filename}{Name of binary dosage file to create}

\item{format}{Primary format of binary dosage file}

\item{subformat}{Secondary format of binary dosage file}
}
\value{
0 success
1 failure
}
\description{
Function to convert and Impute 2 style file to a binary dosage file
}
