% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PlinkBinaryInfo}
\alias{PlinkBinaryInfo}
\title{Function to produce summary of a plink binary genetic file used by GxEScan}
\usage{
PlinkBinaryInfo(geneticFile, mapFile, familyFile)
}
\arguments{
\item{geneticFile}{Name of file with genetic data, normally ends with .bed}

\item{mapFile}{Name of map file associated with genetic data file, normally ends with .bim}

\item{familyFile}{Name of family file associated with genetic data file, normally ends with .fam}
}
\value{
List with data needed by GxEScan
}
\description{
Function to produce summary of a plink binary genetic file used by GxEScan
}
