% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simSubjectData}
\alias{simSubjectData}
\title{Simulated subject data set}
\format{A data frame with 100 rows and 4 variables
\describe{
  \item{iid}{subject id}
  \item{y}{outcome variable}
  \item{x1}{first covariate}
  \item{x2}{second covariate}
}}
\source{
Writers of GxEScanR
}
\usage{
simSubjectData
}
\description{
A data set with subject IDs, an outcome, and two covariates
}
\keyword{datasets}
