% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDInfoSubsetSNPs.R
\name{SubsetBDSNPs}
\alias{SubsetBDSNPs}
\title{Function to subset SNPS}
\usage{
SubsetBDSNPs(bdInfo, minMAF = 0, minRsq = 0, useAltFreq = TRUE)
}
\arguments{
\item{bdInfo}{A binary dosage info list returned from GetBinaryDosageInfo
This information has to be from a binary dosage file in
format 4 or greatef.}

\item{minMAF}{Mimimum minor allele frequency of the SNP. A value of zero
indicates to not subset based on minor allele frequency.
Default value = 0}

\item{minRsq}{Minimum r squared of the imputation for the SNP. A value of
zero indicates to not subset based on r squared value.
Default value = 0}

\item{useAltFreq}{If TRUE, minor allele frequency is calculated from the alternate
allele frequency. If FALSE, the minor allele frequency is used.
Default value is TRUE.}
}
\value{
Vector of SNP indices to use in the imputation - Usually passed
to GxEScan as the snps value
}
\description{
Function to subset SNPs by minor allele frequency and r squared
from a binary dosage file in format 4 or greater. If the binary
dosage file contains more than one group of imputation results,
the minimum r squared and minor allele frequency of the groups
are used for comparison with the values supplied.
}
