% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subjects.R
\name{TestSubjectDataR}
\alias{TestSubjectDataR}
\title{Function to test format of subject data table passed to GxEScan}
\usage{
TestSubjectDataR(subjectData)
}
\arguments{
\item{subjectData}{Data frame with subject data
First column must be a character string
If second column is a character string, the first column is the
is the family ID and the second column is the subject ID otherwise
the first column is the subject
The column following the subject ID is the subject's phenotype
this column must be a numeric value and not an integer
The remian columns are the covariates values. These must be numeric
values. The covariate in the last column is the covariate interacting
with the gene.}
}
\value{
res - List contain the following values
usesFID -- TRUE data frame has family ID, FALSE otherwise
subject -- string matrix with family and subject IDs
phenotype -- numeric vector containing subject phenotypes
covariates -- numberic matrix with covariate values
}
\description{
Function to test format of subject data table passed to GxEScan
}
