% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GxEScanC}
\alias{GxEScanC}
\title{Function to fit models scanning over genotypes}
\usage{
GxEScanC(subjectData, geneticInfo, outputFilename, skippedFilename, minMaf)
}
\arguments{
\item{subjectData}{List returned from SubsetSubjects}

\item{geneticInfo}{List returned from one of the functions to get the required information
about the source of genetic data.}

\item{outputFilename}{Name of output file}

\item{skippedFilename}{Name of file to write info about SNPs that were skipped. If this is blank
no file is written. If this is the same as outputFilename, the skipped SNPs
are written to the output file along with NA for all tests.}

\item{minMaf}{Minimum minor allele frequency. Must be between 0.0001 and 0.25}
}
\value{
0 success
1 failure
}
\description{
Function to fit selected models over genotypes. Results from
these models are then processed by GxETest to display the results
for these one step tests along with two step tests derived from
the one step tests.
}
