% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxEScanR.R
\name{rungweis}
\alias{rungweis}
\title{Routine to run a GWEIS}
\usage{
rungweis(gweismem, bdinfo, snps, outfilename, maf)
}
\arguments{
\item{gweismem}{Models and memory allocated by gweis-mem to run the GWEIS.}

\item{bdinfo}{Information about a Binary Dosage file that contains the
genetic data to run the GWEIS}

\item{snps}{List of SNPs in the Binary Dosage file to perform the GWEIS on.}

\item{outfilename}{Name of the file to contain the output}

\item{maf}{Minimum minor allele frequency of SNPs needed to run test on.}
}
\value{
None
}
\description{
Routine to run a GWEIS
}
