% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subjects.R
\name{TestSubjectDataR}
\alias{TestSubjectDataR}
\title{Function to test format of subject data table passed to GxEScan}
\usage{
TestSubjectDataR()
}
\arguments{
\item{subjectData}{Data frame with subject data
First column must be a character string
If second column is a character string, the first column is the
is the family ID and the second column is the subject ID otherwise
the first column is the subject
The column following the subject ID is the subject's phenotype
this column must be a numeric value and not an integer
The remian columns are the covariates values. These must be numeric
values. The covariate in the last column is the covariate interacting
with the gene.}
}
\value{
0 success
1 failure
}
\description{
Function to test format of subject data table passed to GxEScan
}
